InlineFormLabel => InlineLabel
