--- packages/grafana-ui/src/components/Forms/InlineFormLabel.mdx
@@ -1,15 +1,0 @@
-import { Props } from "@storybook/addon-docs/blocks";
-import { InlineFormLabel } from "./InlineFormLabel";
-
-# InlineFormLabel
-
-A horizontal variant of Label, primarily used in query editors.
-
-# Usage
-```jsx
-<InlineFormLabel width="auto" tooltip="Tooltip content">
-  Simple label
-</InlineFormLabel>
-```
-
-<Props of={InlineFormLabel}/>
