+++ packages/grafana-ui/src/components/Forms/InlineLabel.mdx
@@ -1,0 +1,15 @@
+import { Props } from "@storybook/addon-docs/blocks";
+import { InlineLabel } from "./InlineLabel";
+
+# InlineLabel
+
+A horizontal variant of Label, primarily used in query editors.
+
+# Usage
+```jsx
+<InlineLabel width="auto" tooltip="Tooltip content">
+  Simple label
+</InlineLabel>
+```
+
+<Props of={InlineLabel}/>
